+++ shiny/ui/css_unit/_css_unit.py
@@ -6,6 +6,7 @@ __all__ = (
     "CssUnit",
     "as_css_unit",
     "as_css_padding",
+    "as_width_unit",
 )
 
 CssUnit = Union[
@@ -13,6 +14,9 @@ CssUnit = Union[
     float,
     str,
 ]
+"""
+Possible python types that can be converted into a CSS unit. Numeric values will be converted to pixels. Values equal to `0` will be converted to `"0"`. Strings will be passed through as-is.
+"""
 
 
 @overload
@@ -26,6 +30,19 @@ def as_css_unit(value: CssUnit) -> str:
 
 
 def as_css_unit(value: None | CssUnit) -> None | str:
+    """
+    Convert a value into a CSS unit.
+
+    Parameters
+    ----------
+    value
+        A value to convert into a CSS unit.
+
+    Returns
+    -------
+    :
+        If the `value` is `None`, then `None`. If the value is `0`, then `"0"`. If the `value` is numeric, then a formatted pixel value. Otherwise, the `value` as-is.
+    """
     # TODO-future: Actually validate. Or don't validate, but then change
     # the function name to to_css_unit() or something.
     if isinstance(value, (float, int)):
@@ -48,6 +65,19 @@ def as_css_padding(padding: None) -> None:
 
 
 def as_css_padding(padding: CssUnit | list[CssUnit] | None) -> str | None:
+    """
+    Convert a CSS unit or list of CSS units into a CSS padding value.
+
+    Parameters
+    ----------
+    padding
+        A CSS unit or list of CSS units.
+
+    Returns
+    -------
+    :
+        A CSS padding value.
+    """
     if padding is None:
         return None
 
@@ -58,7 +88,7 @@ def as_css_padding(padding: CssUnit | list[CssUnit] | None) -> str | None:
 
 
 # It seems to be to use % over fr here since there is no gap on the grid
-def to_width_unit(x: str | float | int) -> str:
+def as_width_unit(x: str | float | int) -> str:
     if isinstance(x, (int, float)):
         return as_css_unit(x)
